pyproject.toml
[tool.poetry]
name = "bnb-host-tools"
version = "0.0.2"
description = "A set of tools for managing BnB properties efficiently."
authors = ["Marc <128506536+MarcChen@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "2.100.0"
google-auth = "2.23.4"
google-auth-oauthlib = "1.0.0"
google-auth-httplib2 = "0.1.0"
beautifulsoup4 = "4.12.2"
lxml = "4.9.3"
python-dateutil = "2.8.2"
requests = "2.31.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
